+++ changelog.md
@@ -2,6 +2,7 @@
 - <c-dddddd>New visual identity</c>
 - <c-dddddd>Updated mod information to include everything that is missing</c>
 - <c-dddddd>Move out of Beta</c>
+- <c-dddddd>Fix edit tab sorting with Scrollable Toolbar enabled</c>
 
 # 1.0.0-beta.42
 - <c-dddddd>Improve settings performance</c>